phase_1/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
anthropic
python-multipart
pydantic
pydantic-settings
python-dotenv
